brcmfmac: common: Set MAC address from platform information
On Device Tree platforms, it is customary to be able to set the MAC
address via the Device Tree, as it is often stored in system firmware.
This is particularly relevant for Apple ARM64 platforms, where this
information comes from system configuration and passed through by the
bootloader to the DT.

Implement support for this by fetching the platform MAC address and
using it to *set* the dongle MAC on initialization, instead of getting
it.

Signed-off-by: Hector Martin <marcan@marcan.st>