usb: dwc3: Try usb-role-switch first in dwc3_drd_init
If all of the following are true:
  - the dwc3 controller is setup with usb-role-switch
  - the dwc3 node is connected to a PHY
  - the PHY is connected to a port (e.g. the SS data lane)
  - the PHY has no extcon
dwc3_drd_init will first try to find an extcon node, assume that one
exists in the PHY because it has a port, try to use
extcon_find_edev_by_node and then fail with EPROBE_DEFER because there
can never be an extcon for this PHY node.
The correct usb-role-switch path is never reached. Instead, first check
if usb-role-switch is to be used and only then look for an extcon
device.

Signed-off-by: Sven Peter <sven@svenpeter.dev>