mailbox: sun6i: Unexport broken and useless sun6i_msgbox_peek_data
This op was ambiguously specified, and the way it was interpreted for
this implementation is useless. It has no users, so remove it. The
function is used internally by the driver (for a useful purpose),
so just remove it from the ops structure, which is the context it is
useless in.

Signed-off-by: Hector Martin <marcan@marcan.st>