mailbox: Rename peek_data to poll_data and fix documentation
peek_data was poorly specified and the documentation implied it should
just check for whether there are messages available to be received.
The intent is for this function to be usable in atomic context. Just
checking for available data is completely useless; the subsystem needs
to provide some way to poll for (and actually receive) data in atomic
context for this to be useful.

As you'd expect, all existing drivers that implemented the "peek"
semantics had zero users, since that interpretation of the feature has
zero usefulness. There is only one driver implementing peek_data as
a proper poll: flexrm. That also happens to be the only one with a
user.

So, rename peek_data to poll_data and fix the documentation to actually
describe it properly as a synchronous poll for data. Previous patches
already removed the useless implementations, so this only has to fix the
bcm driver and client besides the subsystem.

Signed-off-by: Hector Martin <marcan@marcan.st>